refact: User 앱 profile 조회/수정 기능 Mypage 앱으로 이동
