+++ user/urls.py
@@ -10,6 +10,4 @@ urlpatterns = [
     path('login/refresh/', TokenRefreshView.as_view(), name='user_token_refresh'),
     path('delete/', views.UserDeleteView.as_view(), name='user_delete'),
 
-    path('profile/',views.UserProfileView.as_view(), name='user_profile')
-
 ]
