+++ user/views.py
@@ -28,18 +28,3 @@ class UserDeleteView(APIView):
             user.delete()
             return Response({"message": "계정이 삭제되었습니다."}, status=status.HTTP_200_OK)
         return Response({"message": "잘못된 요청입니다. 다시 시도해주세요."}, status=status.HTTP_400_BAD_REQUEST)
-
-# 프로필 조회 및 수정
-class UserProfileView(APIView):
-    def get(self, request):
-        profile = Users.objects.get(id=request.user.id)
-        serializer = UserProfileSerializer(profile)
-        return Response(serializer.data)
-    def put(self,request):
-        profile=Users.objects.get(id=request.user.id)
-        profile_serializer = UserProfileSerializer(profile, data=request.data, partial=True)
-        if profile_serializer.is_valid(raise_exception=True):
-            profile_serializer.save()
-            return Response({"data":profile_serializer.data,"message":"프로필 수정 완료"}, status=status.HTTP_201_CREATED)
-        return Response({"message":"저장되지 않았습니다. 다시 시도해주세요"}, status=status.HTTP_400_BAD_REQUEST)
-
